+++ src/pages/farmer/IndexFarmer.tsx
@@ -57,6 +57,7 @@ import { image } from "../../resource";
 import { CheckboxValueType } from "antd/lib/checkbox/Group";
 import { STATUS_NORMAL_MAPPING } from "../../definitions/Status";
 import { ListCheck } from "../../components/dropdownCheck/ListStatusAppType";
+import { DropdownStatus } from "../../components/dropdownCheck/DropDownStatus";
 
 interface SearchSelectType {
   label: any;
@@ -84,14 +85,12 @@ function IndexFarmer() {
   const [sumPlotCard, setSumPlotCard] = useState<any>({
     card1: "รอตรวจสอบ",
     card2: "ไม่อนุมัติ",
+    card3: "ข้อมูลไม่ครบถ้วน",
   });
   const [mainStatus, setMainStatus] = useState<any>("PENDING");
   const [waitPendingDate, setWaitPendingDate] = useState<any>();
   const [appType, setAppType] = useState<any>();
   const [summary, setSummary] = useState<any>();
-  const [visibleStatus, setVisibleStatus] = useState(false);
-  const [indeterminatePending, setIndeterminatePending] = useState(false);
-  const [checkAllPending, setCheckAllPending] = useState(false);
   const [statusArr, setStatusArr] = useState<string[]>([]);
   const [statusArrMain, setStatusArrMain] = useState<string[]>([]);
   const [sortDirection1, setSortDirection1] = useState<string | undefined>(
@@ -165,128 +164,39 @@ function IndexFarmer() {
   const onChangePage = (page: number) => {
     setCurrent(page);
   };
-  const onCheckAllPending = (e: any) => {
-    let value = e.target.value;
-    let checked = e.target.checked;
-    let arr: any = 0;
-    if (checked === true) {
-      arr = [...statusArr, value];
-      setStatusArr([...statusArr, value]);
-      setSearchStatus(value);
-    } else {
-      let d: string[] = statusArr.filter((x) => x != value);
-      arr = [...d];
-      setStatusArr(d);
-      if (d.length == 0) {
-        arr = undefined;
+
+  const onSearchStatus = (value: string, checked: boolean) => {
+    if (statusListPend.map((x) => x).find((c) => c === value)) {
+      let arr: any = 0;
+      if (checked === true) {
+        arr = [...statusArr, value];
+        setStatusArr([...statusArr, value]);
+        setWaitPendingDate(value);
+      } else {
+        let d: string[] = statusArr.filter((x) => x !== value);
+        arr = [...d];
+        setStatusArr(d);
+        if (d.length === 0) {
+          arr = undefined;
+        }
       }
-    }
-    setSearchStatus(arr);
-    setWaitPendingDate(e.target.checked ? statusListPend : []);
-    setIndeterminatePending(false);
-    setCheckAllPending(e.target.checked);
-  };
-  const onChangeListPending = (list: CheckboxValueType[]) => {
-    setSearchStatus(undefined);
-    let arr: any = 0;
-    arr = [...list];
-    setWaitPendingDate(list);
-    setIndeterminatePending(
-      !!list.length && list.length < statusListPend.length
-    );
-    setCheckAllPending(list.length === statusListPend.length);
-  };
-  const onSearchStatus = (e: any) => {
-    let value = e.target.value;
-    let checked = e.target.checked;
-    let arr: any = 0;
-    if (checked === true) {
-      arr = [...statusArrMain, value];
-      setStatusArrMain([...statusArrMain, value]);
-      setSearchStatus(value);
+      setWaitPendingDate(arr);
     } else {
-      let d: string[] = statusArrMain.filter((x) => x != value);
-      arr = [...d];
-      setStatusArrMain(d);
-      if (d.length == 0) {
-        arr = undefined;
+      let arr: any = 0;
+      if (checked === true) {
+        arr = [...statusArrMain, value];
+        setStatusArrMain([...statusArrMain, value]);
+        setSearchStatus(value);
+      } else {
+        let d: string[] = statusArrMain.filter((x) => x !== value);
+        arr = [...d];
+        setStatusArrMain(d);
+        if (d.length === 0) {
+          arr = undefined;
+        }
       }
+      setSearchStatus(arr);
     }
-    setSearchStatus(arr);
-  };
-  const SubStatus = (
-    <Menu
-      items={[
-        mainStatus === "PENDING"
-          ? {
-              label: (
-                <>
-                  <Checkbox
-                    indeterminate={indeterminatePending}
-                    onChange={onCheckAllPending}
-                    checked={checkAllPending}
-                    value="PENDING"
-                  >
-                    รอตรวจสอบ
-                  </Checkbox>
-                  <br />
-                  <Checkbox.Group
-                    value={waitPendingDate}
-                    style={{ width: "100%" }}
-                    onChange={onChangeListPending}
-                  >
-                    <Checkbox style={{ marginLeft: "20px" }} value="FIRST">
-                      1-2 วัน
-                    </Checkbox>
-                    <br />
-                    <Checkbox style={{ marginLeft: "20px" }} value="SECOND">
-                      3-6 วัน
-                    </Checkbox>
-                    <br />
-                    <Checkbox style={{ marginLeft: "20px" }} value="THIRD">
-                      7 วันขึ้นไป
-                    </Checkbox>
-                  </Checkbox.Group>
-                </>
-              ),
-              key: "1",
-            }
-          : {
-              label: (
-                <>
-                  <Checkbox onClick={onSearchStatus} value="ACTIVE">
-                    ใช้งาน
-                  </Checkbox>
-                </>
-              ),
-              key: "1",
-            },
-        mainStatus === "PENDING"
-          ? {
-              label: (
-                <>
-                  <Checkbox onClick={onSearchStatus} value="REJECTED">
-                    ไม่อนุมัติ
-                  </Checkbox>
-                </>
-              ),
-              key: "2",
-            }
-          : {
-              label: (
-                <>
-                  <Checkbox onClick={onSearchStatus} value="INACTIVE">
-                    ปิดการใช้งาน
-                  </Checkbox>
-                </>
-              ),
-              key: "2",
-            },
-      ]}
-    />
-  );
-  const handleVisibleStatus = (newVisible: any) => {
-    setVisibleStatus(newVisible);
   };
 
   const CheckStatus = (e: any) => {
@@ -296,6 +206,7 @@ function IndexFarmer() {
       setSumPlotCard({
         card1: "รอตรวจสอบ",
         card2: "ไม่อนุมัติ",
+        card3: "ข้อมูลไม่ครบถ้วน",
       });
     } else {
       setSearchStatus(undefined);
@@ -339,7 +250,7 @@ function IndexFarmer() {
       setAppTypeArr([...appTypeArr, value]);
       setAppType(value);
     } else {
-      let d: string[] = appTypeArr.filter((x) => x != value);
+      let d: string[] = appTypeArr.filter((x) => x !== value);
       arr = [...d];
       setAppTypeArr(d);
       if (d.length === 0) {
@@ -386,13 +297,19 @@ function IndexFarmer() {
         </div>
       </div>
       <StatusPlots
+        checkPage="DronerPage"
         title1={sumPlotCard?.card1}
         title2={sumPlotCard?.card2}
+        title3={sumPlotCard?.card3}
+        status={mainStatus}
         bgColor1={
           sumPlotCard?.card1 === "รอตรวจสอบ" ? color.Warning : color.Success
         }
         bgColor2={
           sumPlotCard?.card2 === "ไม่อนุมัติ" ? color.Grey : color.Error
+        }
+        bgColor3={
+          sumPlotCard?.card3 === "ข้อมูลไม่ครบถ้วน" ? color.Disable : "none"
         }
         countPlot1={
           mainStatus === "PENDING"
@@ -403,6 +320,11 @@ function IndexFarmer() {
           mainStatus === "ACTIVE"
             ? numberWithCommas(summary?.count_inactive) + " คน"
             : numberWithCommas(summary?.count_reject) + " คน"
+        }
+        countPlot3={
+          mainStatus === "PENDING"
+            ? numberWithCommas(summary?.count_open) + " คน"
+            : null
         }
       />
       <div
@@ -504,19 +426,16 @@ function IndexFarmer() {
             menu="FARMER"
           />
         </div>
-        <div className="col-lg pt-1 p-1">
-          <Dropdown
-            overlay={SubStatus}
-            trigger={["click"]}
-            className="col-lg-12 p-1"
-            onVisibleChange={handleVisibleStatus}
-            visible={visibleStatus}
-          >
-            <Button style={{ color: color.Disable }}>
-              เลือกสถานะ
-              <DownOutlined style={{ verticalAlign: 2 }} />
-            </Button>
-          </Dropdown>
+        <div className="col-lg">
+          <DropdownStatus
+            title="เลือกสถานะ"
+            mainStatus={mainStatus}
+            onSearchType={(value: any, checked: any) =>
+              onSearchStatus(value, checked)
+            }
+            list={searchStatus}
+            menu="DRONER"
+          />
         </div>
         <div className="pt-1">
           <Button